backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
scikit-learn
torch
numpy
pandas
pillow
requests
ultralytics
opencv-python-headless
pyjwt
deep-translator
indic-transliteration
langdetect
